wdtps: Fixed MISRA-C rule 10.3 violations
This patch fixes following MISRAC violations
misra_c_2012_rule_10_3_violation:Implicit conversion from
essential type to different or narrower essential type.


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/wdtps/src/xwdtps_selftest.c
@@ -107,7 +107,7 @@ s32 XWdtPs_SelfTest(XWdtPs *InstancePtr)
 		XWdtPs_WriteReg(InstancePtr->Config.BaseAddress,
 				  XWDTPS_ZMR_OFFSET,
 				  (ZmrOrig | (u32)XWDTPS_ZMR_ZKEY_VAL));
-		Status = XST_FAILURE;
+		Status = (s32)XST_FAILURE;
 	} else {
 
 
@@ -128,7 +128,7 @@ s32 XWdtPs_SelfTest(XWdtPs *InstancePtr)
 			XWdtPs_WriteReg(InstancePtr->Config.BaseAddress,
 					  XWDTPS_ZMR_OFFSET,
 					  ZmrOrig | XWDTPS_ZMR_ZKEY_VAL);
-			Status = XST_FAILURE;
+			Status = (s32)XST_FAILURE;
 
 		} else {
 
@@ -139,7 +139,7 @@ s32 XWdtPs_SelfTest(XWdtPs *InstancePtr)
 			XWdtPs_WriteReg(InstancePtr->Config.BaseAddress, XWDTPS_ZMR_OFFSET,
 					  ZmrOrig | XWDTPS_ZMR_ZKEY_VAL);
 
-			Status = XST_SUCCESS;
+			Status = (s32)XST_SUCCESS;
 		}
 	}
 	return Status;
